chore: move stuff to correct namespace
