+++ Modules/Tokens/test/Tokens.Domain.Tests/Tests/Tokens/TokenTests.cs
@@ -1,5 +1,5 @@
 ﻿using Backbone.BuildingBlocks.Domain.Exceptions;
-using Backbone.Modules.Tokens.Domain.DomainEvents;
+using Backbone.Modules.Tokens.Domain.DomainEvents.Outgoing;
 using Backbone.Modules.Tokens.Domain.Entities;
 using Backbone.Modules.Tokens.Domain.Tests.TestHelpers;
